backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Caching
redis==5.0.1

# GCP
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
google-auth==2.23.4

# LLM Integration
langchain==0.1.0
openai==1.3.7
anthropic==0.7.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-multipart==0.0.6

# Playwright for test execution
playwright==1.40.0 